requirements.txt
flake8==3.9.2
pandas==1.5.0
numpy==1.23.0
seaborn==0.13.2
pyyaml==6.0.0
pyarrow==8.0.0
pytest==6.2.4
duckdb==0.5.1
joblib==1.0.1
matplotlib==3.5.2
torch==2.2.1
ray==2.2.0
hyperopt==0.2.7
requests==2.26.0
tabulate==0.8.9
tensorboardX==2.4
psycopg2==2.9.5
scikit-learn==1.1.1
neuralforecast==1.6.4
